Part 1 of ???: Fixing annotations on unicode vs ansi
so i didnt realize that technically the union { was pointless and made it look like the description was < CAPI provider name

This is yet another thing with the duality support for Unicode/ANSI, and this is me fixing a previous mistake. Each struct appears to have
a field for Unicode and ANSI. I wonder if this bogs up resources.